+++ x/pocketcore/types/cache.go
@@ -263,7 +263,7 @@ func SessionIterator() SessionIt {
 }
 
 // "GetEvidence" - Retrieves the evidence object from the storage
-func GetEvidence(header SessionHeader, evidenceType EvidenceType, max sdk.Int) (evidence Evidence, err error) {
+func GetEvidence(header SessionHeader, evidenceType EvidenceType, max sdk.BigInt) (evidence Evidence, err error) {
 	// generate the key for the evidence
 	key, err := KeyForEvidence(header, evidenceType)
 	if err != nil {
@@ -391,7 +391,7 @@ func GetProof(header SessionHeader, evidenceType EvidenceType, index int64) Proo
 }
 
 // "SetProof" - Sets a proof object in the evidence, using the header and evidence type
-func SetProof(header SessionHeader, evidenceType EvidenceType, p Proof, max sdk.Int) {
+func SetProof(header SessionHeader, evidenceType EvidenceType, p Proof, max sdk.BigInt) {
 	// retireve the evidence
 	evidence, err := GetEvidence(header, evidenceType, max)
 	// if not found generate the evidence object
@@ -409,7 +409,7 @@ func IsUniqueProof(p Proof, evidence Evidence) bool {
 }
 
 // "GetTotalProofs" - Returns the total number of proofs for a piece of evidence
-func GetTotalProofs(h SessionHeader, et EvidenceType, maxPossibleRelays sdk.Int) (Evidence, int64) {
+func GetTotalProofs(h SessionHeader, et EvidenceType, maxPossibleRelays sdk.BigInt) (Evidence, int64) {
 	// retrieve the evidence
 	evidence, err := GetEvidence(h, et, maxPossibleRelays)
 	if err != nil {
